ouch, more size_t pain discovered by Sergey
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@22516 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/fitsbin.c
@@ -508,7 +508,7 @@ static int read_chunk(fitsbin_t* fb, fitsbin_chunk_t* chunk) {
 		int i;
 		chunk->data = malloc(expected);
 		for (i=0; i<chunk->nrows; i++) {
-			memcpy(((char*)chunk->data) + i * chunk->itemsize,
+			memcpy(((char*)chunk->data) + (size_t)i * (size_t)chunk->itemsize,
 				   bl_access(inmemext->items, i), chunk->itemsize);
 		}
 		// delete inmemext->items ?
